+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## Unreleased
+
+### Bugfix
+
+- Fix Bluehawk Upgrade Tech not applying to the anti-orbital portion of the weapon
+
 ## v5.80.0 - 2025-07-04
 
 ### Added